% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_check.R
\name{na_check}
\alias{na_check}
\title{na checks}
\usage{
na_check(
  x,
  na_type = c(),
  na_consecutive_n = NULL,
  na_max_n = NULL,
  na_max_prop = NULL,
  na_min_n = NULL,
  na_FUN = NULL,
  ...
)
}
\arguments{
\item{x}{A vector for which missing values checks need to be performed.}

\item{na_type}{A character vector specifying the types of missing values checks to perform. Available options are "n" (count of missing values), "prop" (proportion of missing values), "n_non_miss" (count of non-missing values), "FUN" (custom function), and "con" (consecutive missing values).}

\item{na_consecutive_n}{Maximum number of consecutive missing values allowed (used with "con" type).}

\item{na_max_n}{Maximum count of missing values allowed (used with "n" type).}

\item{na_max_prop}{Maximum proportion of missing values allowed (used with "prop" type).}

\item{na_min_n}{Minimum count of non-missing values required (used with "n_non_miss" type).}

\item{na_FUN}{A custom function to determine missing values criteria (used with "FUN" type).}

\item{...}{Additional arguments to be passed to the custom function {na_FUN}.}
}
\value{
{TRUE} if all specified missing values checks pass, {FALSE} otherwise.
}
\description{
This function performs various types of missing values checks on a given vector \code{x} based on specified criteria.
}
\examples{

# Example 1: Using a custom function for missing values check
custom_check <- function(x) { sum(is.na(x)) == 0 }
data <- c(1, 2, 3, 4, 5)
na_check(data, na_type = "FUN", na_FUN = custom_check)

# Example 2: Checking consecutive missing values
data <- c(1, 2, NA, NA, 5, NA)
na_check(data, na_type = "con", na_consecutive_n = 2)
}
