% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ang_dev_circular.R
\name{summary_ang_dev_circular}
\alias{summary_ang_dev_circular}
\title{Angular Deviation}
\usage{
summary_ang_dev_circular(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{Logical indicating whether to remove missing values (default is FALSE).}

\item{na_type}{Character specifying the type of missing values to consider if na.rm is TRUE.}

\item{...}{Additional arguments to be passed to circular::angular.deviation() function.}
}
\value{
The angular deviation of the input data.
}
\description{
This function calculates the angular deviation of a circular data vector.
}
\examples{
# Example usage:
data <- circular::circular(c(0, pi/2, pi, 3*pi/2, 2*pi))
summary_ang_dev_circular(data)


# summary_ang_dev_circular(data, na.rm = TRUE, na_type = "na.fail")


# summary_ang_dev_circular(data, na.rm = TRUE, na_type = "na.exclude")
 

}
