% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ang_var_circular.R
\name{summary_ang_var_circular}
\alias{summary_ang_var_circular}
\title{Angular Variance of Circular Data}
\usage{
summary_ang_var_circular(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A vector of circular data, in radians.}

\item{na.rm}{A logical value indicating whether to remove NA values from the input vector before calculation.}

\item{na_type}{A character string specifying how NA values should be treated. Options include "omit", "na.fail", or "propagate".}

\item{...}{Additional arguments to be passed to the circular::angular.variance function.}
}
\value{
The angular variance of the input circular data.
}
\description{
This function calculates the angular variance of circular data.
}
\examples{
x <- circular::circular(c(0.1, 1.2, 2.5, NA, 3.8))
summary_ang_var_circular(x) # Calculates angular variance without removing NAs
summary_ang_var_circular(x, na.rm = TRUE) # Calculates angular variance after removing NAs
#summary_ang_var_circular(x, na.rm = TRUE, 
}
