% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_count.R
\name{summary_count}
\alias{summary_count}
\title{Summary Count Function}
\usage{
summary_count(x, ...)
}
\arguments{
\item{x}{A vector or list.}

\item{...}{Additional arguments (not used in this function).}
}
\value{
The length of the input vector or list.
}
\description{
This function calculates the length of a vector or list.
}
\examples{
x <- c(1, 2, 3, 4, 5)
summary_count(x) 
}
