% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_count_non_missing.R
\name{summary_count_non_missing}
\alias{summary_count_non_missing}
\title{Summary count non missing}
\usage{
summary_count_non_missing(x, ...)
}
\arguments{
\item{x}{The input vector for which you want to count non-missing values.}

\item{...}{Additional arguments (not named) that can be passed to the function.}
}
\value{
The count of non-missing (non-NA) values in the input vector \code{x}.
}
\description{
Summary count non missing
}
\examples{
summary_count_non_missing(c(1, 2, NA, 4, NA, 6))
# Returns: 4
}
