% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_max.R
\name{summary_max}
\alias{summary_max}
\title{Summary Maximum Value Function}
\usage{
summary_max(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector for which you want to find the maximum value.}

\item{na.rm}{Logical, indicating whether missing values should be removed (default is FALSE).}

\item{na_type}{Character string specifying how to handle missing values. Possible values include "" (default), "na.omit", "na.exclude", "na.fail".}

\item{...}{Additional arguments (not used in this function).}
}
\value{
The maximum value in the input vector 'x'.
}
\description{
This function calculates the maximum value in a numeric vector 'x'.
}
\examples{
x <- c(1, 2, 3, 4, 5)
summary_max(x) # Returns the maximum value in 'x'

# Example with missing values
#y <- c(1, 2, NA, 4, 5)
#summary_max(y, na.rm = TRUE, na_type = "na.omit") # Returns the maximum value with missing values removed
}
