% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_max_circular.R
\name{summary_max_circular}
\alias{summary_max_circular}
\title{Calculate Maximum of Circular Data}
\usage{
summary_max_circular(
  x,
  na.rm = FALSE,
  names = FALSE,
  type = 7,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector containing circular data for which the maximum needs to be calculated.}

\item{na.rm}{A logical value indicating whether missing values (\code{NA}s) should be removed before calculation.}

\item{names}{A logical value indicating whether the result should be named.}

\item{type}{An integer indicating the type of quantile computation. Options include 1 to 9.}

\item{na_type}{A character string specifying the type of missing values check to perform. Options include "n", "prop", "n_non_miss", "FUN", and "con".}

\item{...}{Additional arguments to be passed to the circular::quantile.circular function.}
}
\value{
The maximum of the input circular data vector \code{x}. If \code{na.rm} is \code{TRUE}, missing values are removed before calculation. If the data is empty or all missing (\code{NA}) values, the result will be \code{NA}.
}
\description{
This function calculates the maximum of a circular data vector \code{x}.
}
\examples{
# Example 1: Calculate maximum without removing missing values
data <- circular::circular(c(0, pi/2, pi, 3*pi/2))
summary_max_circular(data)

# Example 2: Calculate maximum after removing missing values
data <- circular::circular(c(0, pi/2, NA, 3*pi/2))
summary_max_circular(data, na.rm = TRUE)

}
