% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mean.R
\name{summary_mean}
\alias{summary_mean}
\title{Summary  Mean}
\usage{
summary_mean(
  x,
  add_cols,
  weights = NULL,
  na.rm = FALSE,
  trim = 0,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{add_cols}{Deprecated parameter; no longer in use.}

\item{weights}{Optional numeric vector of weights for weighted mean calculations (default is NULL).}

\item{na.rm}{Logical. Should NA values be removed from the input vector \code{x} before calculation (default is FALSE)?}

\item{trim}{Proportion (0 to 0.5) of data to be trimmed from both ends before calculating the mean (default is 0).}

\item{na_type}{A character string specifying how to handle NAs if \code{na.rm} is TRUE (default is "").}

\item{...}{Additional arguments (not used in this function).}
}
\value{
The mean value of the input numeric vector \code{x}.
}
\description{
This function calculates the mean of a numeric vector, either a simple mean or a weighted mean.
}
\examples{
# Example 1: Calculate the mean of a numeric vector
data <- c(1, 2, 3, 4, 5)
summary_mean(data)

# Example 2: Calculate the mean with NA removal
data <- c(1, 2, NA, 4, 5)
summary_mean(data, na.rm = TRUE)

# Example 3: Calculate the weighted mean
data <- c(1, 2, 3, 4, 5)
weights <- c(0.1, 0.2, 0.3, 0.2, 0.2)
summary_mean(data, weights = weights)

# Example 4: Calculate the trimmed mean
data <- c(1, 2, 3, 4, 5)
summary_mean(data, trim = 0.2)
}
