% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_median.R
\name{summary_median}
\alias{summary_median}
\title{Summary Median Function}
\usage{
summary_median(x, na.rm = FALSE, weights = NULL, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector for which you want to calculate the median.}

\item{na.rm}{Logical, indicating whether missing values should be removed (default is FALSE).}

\item{weights}{A vector of weights to be used in the calculation (default is NULL).}

\item{na_type}{Character string specifying how to handle missing values. Possible values include "" (default), "na.omit", "na.exclude", "na.fail".}

\item{...}{Additional arguments (not used in this function).}
}
\value{
The median value of the input vector 'x'.
}
\description{
This function calculates the median of a numeric vector 'x'.
}
\examples{
x <- c(1, 2, 3, 4, 5)
summary_median(x) # Returns the median of 'x'

# Example with missing values and weights
#y <- c(1, 2, NA, 4, 5)
#w <- c(0.1, 0.2, 0.1, 0.3, 0.2)
# Returns weighted median with missing values removed
#summary_median(y, na.rm = TRUE, weights = w,na_type = "na.omit") 

}
