% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Median_circular.R
\name{summary_median_circular}
\alias{summary_median_circular}
\title{Median of Circular Data}
\usage{
summary_median_circular(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A vector containing circular data for which the median needs to be calculated.}

\item{na.rm}{A logical value indicating whether missing values ({NA}s) should be removed before calculation.}

\item{na_type}{A character string specifying the type of missing values check to perform. Options include "n", "prop", "n_non_miss", "FUN", and "con".}

\item{...}{Additional arguments to be passed to the circular::median.circular function.}
}
\value{
The circular median of the input data vector {x}. If {na.rm} is {TRUE}, missing values are removed before calculation. If the data is empty or all missing ({NA}) values, the result will be {NA}.
}
\description{
This function calculates the median of a circular data vector {x}, taking into account circular nature of the data.
}
\examples{
# Example 1: Calculate circular median without removing missing values
data <- circular::circular(c(0, pi/2, pi, 3*pi/2))
summary_median_circular(data)

# Example 2: Calculate circular median after removing missing values
data <- circular::circular(c(0, pi/2, NA, 3*pi/2))
summary_median_circular(data, na.rm = TRUE)

}
