% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianHL.R
\name{summary_medianHL_circular}
\alias{summary_medianHL_circular}
\title{Calculate High-Low Median of Circular Data}
\usage{
summary_medianHL_circular(
  x,
  na.rm = FALSE,
  method = c("HL1", "HL2", "HL3"),
  prop = NULL,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector containing circular data for which the high-low median needs to be calculated.}

\item{na.rm}{A logical value indicating whether missing values ({NA}s) should be removed before calculation.}

\item{method}{A character vector specifying the method to use for calculating high-low median. Available options are "HL1", "HL2", and "HL3".}

\item{prop}{A numeric value specifying the proportion of data to be considered for calculating high-low median (used with "HL2" and "HL3" methods).}

\item{na_type}{A character string specifying the type of missing values check to perform. Options include "n", "prop", "n_non_miss", "FUN", and "con".}

\item{...}{Additional arguments to be passed to the circular::medianHL.circular function.}
}
\value{
The high-low median of the input circular data vector {x}. If {na.rm} is {TRUE}, missing values are removed before calculation. If the data is empty or all missing ({NA}) values, the result will be {NA}.
}
\description{
This function calculates the high-low median of a circular data vector {x}, considering different methods and a proportion of data.
}
\examples{
# Example 1: Calculate high-low median with method "HL1"
data <- circular::circular(c(0, pi/2, pi, 3*pi/2))
summary_medianHL_circular(data, method = "HL1")

# Example 2: Calculate high-low median with method "HL2"
data <- circular::circular(c(0, pi/2, pi, 3*pi/2))
summary_medianHL_circular(data, method = "HL2", prop = 0.5)

# Example 3: Calculate high-low median with method "HL3"
data <- circular::circular(c(0, pi/2, pi, 3*pi/2))
summary_medianHL_circular(data, method = "HL3", prop = 0.25)

# Example 4: Calculate high-low median after removing missing values
data <- circular::circular(c(0, pi/2, NA, 3*pi/2))
summary_medianHL_circular(data, na.rm = TRUE, method = "HL1")

}
