% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Min_circular.R
\name{summary_min_circular}
\alias{summary_min_circular}
\title{#' Calculate Minimum of Circular Data}
\usage{
summary_min_circular(
  x,
  na.rm = FALSE,
  names = FALSE,
  type = 7,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector containing circular data for which the minimum needs to be calculated.}

\item{na.rm}{A logical value indicating whether missing values ({NA}s) should be removed before calculation.}

\item{names}{A logical value indicating whether the result should be named.}

\item{type}{An integer indicating the type of quantile computation. Options include 1 to 9.}

\item{na_type}{A character string specifying the type of missing values check to perform. Options include "n", "prop", "n_non_miss", "FUN", and "con".}

\item{...}{Additional arguments to be passed to the circular::quantile.circular function.}
}
\value{
The minimum of the input circular data vector {x}. If {na.rm} is {TRUE}, missing values are removed before calculation. If the data is empty or all missing ({NA}) values, the result will be {NA}.
}
\description{
This function calculates the minimum of a circular data vector {x}.
}
\examples{
# Example 1: Calculate minimum without removing missing values
data <- circular::circular(c(0, pi/2, pi, 3*pi/2))
summary_min_circular(data)

# Example 2: Calculate minimum after removing missing values
data <- circular::circular(c(0, pi/2, NA, 3*pi/2))
summary_min_circular(data, na.rm = TRUE)


}
