% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_Q1_circular.R
\name{summary_Q1_circular}
\alias{summary_Q1_circular}
\title{Summary Q1 Circular}
\usage{
summary_Q1_circular(
  x,
  na.rm = FALSE,
  names = FALSE,
  type = 7,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{Logical. Should NA values be removed from the input vector \code{x} before calculation?}

\item{names}{Logical. Should the resulting quantile be named?}

\item{type}{An integer indicating the type of quantile calculation to be performed (default is 7).}

\item{na_type}{A character string specifying how to handle NAs if \code{na.rm} is TRUE (default is "").}

\item{...}{Additional arguments to be passed to the \code{circular::quantile.circular} function.}
}
\value{
The first quartile (Q1) for the circular data.
}
\description{
This function calculates the first quartile (Q1) for circular data using the \code{circular::quantile.circular} function.
}
\examples{
# Example 1: Calculate Q1 for circular data
data <- circular::circular(c(0, 1, 2, 3, 4), type = "angles")
summary_Q1_circular(data)

# Example 2: Calculate Q1 with NA removal
data <- circular::circular(c(0, 1, NA, 3, 4), type = "angles")
summary_Q1_circular(data, na.rm = TRUE)


}
