% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_Q3_circular.R
\name{summary_Q3_circular}
\alias{summary_Q3_circular}
\title{Q3 circular}
\usage{
summary_Q3_circular(
  x,
  na.rm = FALSE,
  names = FALSE,
  type = 7,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{A logical value indicating whether to remove NA values from 'x' before calculation.}

\item{names}{A logical value indicating whether to return names for the quartile values.}

\item{type}{An integer specifying the type of circular data (default is 7).}

\item{na_type}{A character string specifying the type of NA handling (if any).}

\item{...}{Additional arguments to be passed to 'circular::quantile.circular'.}
}
\value{
The third quartile (75th percentile) of the circular data.
}
\description{
This function calculates the third quartile (75th percentile) of circular data.
}
\examples{
#summary_Q3_circular(circular_data)
#summary_Q3_circular(circular_data, na.rm = TRUE)
#summary_Q3_circular(circular_data, type = 3, names = TRUE)
}
