% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_quantile.R
\name{summary_quantile}
\alias{summary_quantile}
\title{Summary Quantile Function}
\usage{
summary_quantile(x, na.rm = FALSE, weights = NULL, probs, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector for which you want to calculate a quantile.}

\item{na.rm}{Logical, indicating whether missing values should be removed (default is FALSE).}

\item{weights}{A vector of weights to be used in the calculation (default is NULL).}

\item{probs}{Numeric vector of probabilities with values between 0 and 1, indicating which quantiles to calculate.}

\item{na_type}{Character string specifying how to handle missing values. Possible values include "" (default), "na.omit", "na.exclude", "na.fail".}

\item{...}{Additional arguments (not used in this function).}
}
\value{
The specified quantile(s) of the input vector 'x'.
}
\description{
This function calculates a quantile of a numeric vector 'x'.
}
\examples{
x <- c(1, 2, 3, 4, 5)
summary_quantile(x, probs = 0.25) # Returns the 25th percentile of 'x'

# Example with missing values and weights
 #y <- c(1, 2, NA, 4, 5)
 #w <- c(0.1, 0.2, 0.1, 0.3, 0.2)
 #summary_quantile(y, na.rm = TRUE,
     weights = w, probs = c(0.25, 0.75), na_type = "na.omit") # Returns weighted quantiles with missing values removed
}
