% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_quantile_circular.R
\name{summary_quantile_circular}
\alias{summary_quantile_circular}
\title{Quantiles of Circular Data}
\usage{
summary_quantile_circular(
  x,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  names = FALSE,
  type = 7,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector of circular data.}

\item{probs}{A numeric vector of probabilities, ranging from 0 to 1, at which to calculate quantiles.}

\item{na.rm}{Logical. Should NA values be removed from the input vector \code{x} before calculation?}

\item{names}{Logical. Should the resulting quantiles be named?}

\item{type}{An integer indicating the type of quantile calculation to be performed (default is 7).}

\item{na_type}{A character string specifying how to handle NAs if \code{na.rm} is TRUE (default is "").}

\item{...}{Additional arguments to be passed to the \code{circular::quantile.circular} function.}
}
\value{
A numeric vector of quantiles corresponding to the specified probabilities.
}
\description{
This function calculates quantiles for circular data using the \code{circular::quantile.circular} function.
}
\examples{
# Example 1: Calculate quantiles for circular data
data <- circular::circular(c(0, 1, 2, 3, 4), type = "angles")
summary_quantile_circular(data, probs = c(0.25, 0.75))

# Example 2: Calculate quantiles with NA removal
data <- circular::circular(c(0, 1, NA, 3, 4), type = "angles")
summary_quantile_circular(data, probs = c(0.25, 0.75), na.rm = TRUE)

}
