% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_range.R
\name{summary_range}
\alias{summary_range}
\title{Summary Range Function}
\usage{
summary_range(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector for which you want to calculate the range.}

\item{na.rm}{Logical, indicating whether missing values should be removed (default is FALSE).}

\item{na_type}{Character string specifying how to handle missing values. Possible values include "" (default), "na.omit", "na.exclude", "na.fail".}

\item{...}{Additional arguments (not used in this function).}
}
\value{
The range (max - min) of the input vector 'x'.
}
\description{
This function calculates the range of a numeric vector 'x', which is the difference between the maximum and minimum values.
}
\examples{
x <- c(1, 2, 3, 4, 5)
summary_range(x) # Returns the range of 'x'

# Example with missing values
#y <- c(1, 2, NA, 4, 5)
#summary_range(y, na.rm = TRUE, na_type = "na.omit") # Returns the range with missing values removed
}
