% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ang_var_circular.R
\name{summary_range_circular}
\alias{summary_range_circular}
\title{Circular Range}
\usage{
summary_range_circular(
  x,
  test = FALSE,
  na.rm = FALSE,
  finite = FALSE,
  control.circular = list(),
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector of circular data in radians.}

\item{test}{A logical value indicating whether to perform a test for uniformity (default is FALSE).}

\item{na.rm}{A logical value indicating whether to remove NA values from the input vector before calculation (default is FALSE).}

\item{finite}{A logical value indicating whether to consider finite values only (default is FALSE).}

\item{control.circular}{A list of control parameters to customize the circular range calculation.}

\item{na_type}{A character string specifying how NA values should be treated. Options include "omit", "na.fail", or "propagate" (default is an empty string).}

\item{...}{Additional arguments to be passed to the circular::range.circular function.}
}
\value{
The circular range of the input circular data, and optionally the result of the uniformity test.
}
\description{
This function calculates the circular range of circular data and optionally performs a test for uniformity.
}
\examples{
x <- circular::circular(c(0.1, 1.2, 2.5, 3.8, 5.0))
summary_range_circular(x) # Calculates the circular range
summary_range_circular(x, test = TRUE) # Calculates the circular range and performs a uniformity test
#summary_range_circular(x, na.rm = TRUE, na_type = "omit") # Calculates the circular range after removing NAs using "omit" method
}
