% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rho_circular.R
\name{summary_rho_circular}
\alias{summary_rho_circular}
\title{Circular Rho}
\usage{
summary_rho_circular(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A vector of circular data, in radians.}

\item{na.rm}{A logical value indicating whether to remove NA values from the input vector before calculation (default is FALSE).}

\item{na_type}{A character string specifying how NA values should be treated. Options include "omit", "na.fail", or "propagate" (default is an empty string).}

\item{...}{Additional arguments to be passed to the circular::rho.circular function.}
}
\value{
The circular rho  of the input circular data.
}
\description{
This function calculates the circular rho (mean resultant length) of circular data.
}
\examples{
x <- circular::circular(c(0.1, 1.2, 2.5, 3.8, 5.0))
summary_rho_circular(x) # Calculates the circular rho
summary_rho_circular(x, na.rm = TRUE) # Calculates the circular rho after removing NAs
summary_rho_circular(x, na_rm = TRUE, na_type = "omit") # Calculates the circular rho after removing NAs using "omit" method
}
