% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sd_circular.R
\name{summary_sd_circular}
\alias{summary_sd_circular}
\title{Summary sd circular}
\usage{
summary_sd_circular(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{Logical indicating whether to remove missing values (default is FALSE).}

\item{na_type}{Character specifying the type of missing values to consider if na.rm is TRUE.}

\item{...}{Additional arguments to be passed to circular::sd.circular() function.}
}
\value{
The circular standard deviation of the input data.
}
\description{
This function calculates the circular standard deviation of a circular data vector.
}
\examples{
# Example usage:
data <- circular::circular(c(0, pi/2, pi, 3*pi/2, 2*pi))
summary_sd_circular(data)
}
