% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_trimmed_mean.R
\name{summary_trimmed_mean}
\alias{summary_trimmed_mean}
\title{Trimmed Mean}
\usage{
summary_trimmed_mean(
  x,
  add_cols,
  weights = NULL,
  na.rm = FALSE,
  trimmed = 0,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{add_cols}{Deprecated parameter; no longer in use.}

\item{weights}{Optional numeric vector of weights for weighted mean calculations (default is NULL).}

\item{na.rm}{Logical. Should NA values be removed from the input vector \code{x} before calculation (default is FALSE)?}

\item{trimmed}{Proportion (0 to 0.5) of data to be trimmed from both ends before calculating the mean (default is 0).}

\item{na_type}{A character string specifying how to handle NAs if \code{na.rm} is TRUE (default is "").}

\item{...}{Additional arguments (not used in this function).}
}
\value{
The trimmed mean value of the input numeric vector \code{x}.
}
\description{
This function calculates the trimmed mean of a numeric vector.
}
\examples{
# Example 1: Calculate the trimmed mean of a numeric vector
data <- c(1, 2, 3, 4, 5)
summary_trimmed_mean(data)

# Example 2: Calculate the trimmed mean with NA removal
data <- c(1, 2, NA, 4, 5)
summary_trimmed_mean(data, na.rm = TRUE)

# Example 3: Calculate the trimmed mean with custom trimming
data <- c(1, 2, 3, 4, 5)
summary_trimmed_mean(data, trimmed = 0.2)
}
