% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_var_circular.R
\name{summary_var_circular}
\alias{summary_var_circular}
\title{Circular Variance}
\usage{
summary_var_circular(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{Logical indicating whether to remove missing values (default is FALSE).}

\item{na_type}{Character specifying the type of missing values to consider if na.rm is TRUE.}

\item{...}{Additional arguments to be passed to circular::var.circular() function.}
}
\value{
The circular variance of the input data.
}
\description{
This function calculates the circular variance of a circular data vector.
}
\examples{
# Example usage:
data <- circular::circular(c(0, pi/2, pi, 3*pi/2, 2*pi))
summary_var_circular(data)
# [1] 0.8

#summary_var_circular(data, na.rm = TRUE, na_type = "na.fail")
# [1] NA

#summary_var_circular(data, na.rm = TRUE, na_type = "na.exclude")
# [1] 0.8

}
